requirements.txt
torch
torchvision~=0.15.2
numpy~=1.22.3
matplotlib~=3.7.0
pillow~=9.4.0
wandb~=0.15.4
ursina~=5.3.0